requirements.txt
python-periphery>=1.0.0
RPi.GPIO>=0.5.11
